fix(calculator): 요구사항에 맞게 소수 입력도 처리 가능하도록 수정
-기존에는 정수만 고려했지만, 요구사항의 '양수'에 맞게 양의 소수도 처리 가능하게 로직을 변경하였습니다.
- toIntOrNull() 메서드를 toDoubleOrNull()로 교체하여 소수점 파싱도  가능하게 하였습니다.
- 또한 파싱 결과를 List<Int>로 주고받던 부분을 List<double>로 수정  하였습니다.

+++ src/main/kotlin/calculator/ConsoleView.kt
@@ -6,5 +6,5 @@ class ConsoleView {
         println("덧셈할 문자열을 입력해 주세요.")
         return readLine()
     }
-    fun printResult(result : Int) = println("결과 : $result")
+    fun printResult(result : Double) = println("결과 : $result")
 }
